+++ src/ContaoKongresseBundle.php
@@ -1,9 +1,22 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of schachbulle/contao-kongresse-bundle.
+ *
+ * (c) Frank Hoppe
+ *
+ * @license LGPL-3.0-or-later
+ */
+
 namespace Schachbulle\ContaoKongresseBundle;
 
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
+/**
+ * Symfony-Bundle-Klasse der Erweiterung "Kongresse & Ausschüsse".
+ */
 class ContaoKongresseBundle extends Bundle
 {
 }