+++ src/Resources/contao/languages/de/tl_module.php
@@ -1,18 +1,7 @@
 <?php
 
 /**
- * Contao Open Source CMS
- *
- * Copyright (c) 2005-2014 Leo Feyer
- *
- * @package   fh-counter
- * @author    Frank Hoppe
- * @license   GNU/LGPL
- * @copyright Frank Hoppe 2014
- */
-
-/**
- * Felder für Ausgabemodul
+ * Deutsche Beschriftungen der Moduleinstellungen in tl_module.
  */
 $GLOBALS['TL_LANG']['tl_module']['kongresse_legend'] = 'Einstellungen';
 $GLOBALS['TL_LANG']['tl_module']['kongresse_from'] = array('Von Jahr', 'Von Jahr (optional)');