+++ src/Resources/contao/languages/en/modules.php
@@ -1,0 +1,8 @@
+<?php
+
+/**
+ * Englische Bezeichnungen des Backend- und des Frontend-Moduls.
+ */
+
+$GLOBALS['TL_LANG']['MOD']['kongresse'] = array('Congresses & committees', 'Manage congresses and committees');
+$GLOBALS['TL_LANG']['FMD']['kongresse'] = array('Congresses & committees', 'List congresses and committees');