+++ src/Resources/contao/languages/en/tl_kongresse.php
@@ -1,0 +1,42 @@
+<?php
+
+/**
+ * Englische Beschriftungen der Tabelle tl_kongresse.
+ */
+
+/**
+ * Buttons für Operationen
+ */
+$GLOBALS['TL_LANG']['tl_kongresse']['new'] = array('New event', 'Create a new event');
+$GLOBALS['TL_LANG']['tl_kongresse']['edit'] = array('Edit event', 'Edit event %s');
+$GLOBALS['TL_LANG']['tl_kongresse']['copy'] = array('Duplicate event', 'Duplicate event %s');
+$GLOBALS['TL_LANG']['tl_kongresse']['delete'] = array('Delete event', 'Delete event %s');
+$GLOBALS['TL_LANG']['tl_kongresse']['toggle'] = array('Publish/unpublish event', 'Publish/unpublish event %s');
+$GLOBALS['TL_LANG']['tl_kongresse']['show'] = array('Event details', 'Show the details of event %s');
+
+/**
+ * Backend-Modul: Übersetzungen im Eingabeformular
+ */
+$GLOBALS['TL_LANG']['tl_kongresse']['congress_legend'] = 'Event';
+$GLOBALS['TL_LANG']['tl_kongresse']['typ'] = array('Type', 'Type of the event');
+$GLOBALS['TL_LANG']['tl_kongresse']['online'] = array('Online', 'Online event');
+$GLOBALS['TL_LANG']['tl_kongresse']['jahr'] = array('Year', 'Year of the event');
+$GLOBALS['TL_LANG']['tl_kongresse']['ort'] = array('Location', 'Location of the event');
+$GLOBALS['TL_LANG']['tl_kongresse']['datum_von'] = array('Start date', 'Start date, or leave empty');
+$GLOBALS['TL_LANG']['tl_kongresse']['datum_bis'] = array('End date', 'End date, or leave empty');
+$GLOBALS['TL_LANG']['tl_kongresse']['info'] = array('Information', 'Additional information to be displayed');
+
+$GLOBALS['TL_LANG']['tl_kongresse']['files_legend'] = 'Downloads and event page';
+$GLOBALS['TL_LANG']['tl_kongresse']['file_broschuere'] = array('Book/brochure', 'Book or brochure of the event');
+$GLOBALS['TL_LANG']['tl_kongresse']['file_protokoll'] = array('Minutes', 'Minutes of the event');
+$GLOBALS['TL_LANG']['tl_kongresse']['url'] = array('Link', 'Link to the event page');
+$GLOBALS['TL_LANG']['tl_kongresse']['newWindow'] = array('New window', 'Open the link in a new window');
+
+$GLOBALS['TL_LANG']['tl_kongresse']['extra_legend'] = 'Additional links';
+$GLOBALS['TL_LANG']['tl_kongresse']['extra_links'] = array('Links', '');
+$GLOBALS['TL_LANG']['tl_kongresse']['extra_links_url'] = array('URL', '');
+$GLOBALS['TL_LANG']['tl_kongresse']['extra_links_text'] = array('Link text (optional)', '');
+$GLOBALS['TL_LANG']['tl_kongresse']['extra_links_newWindow'] = array('New window', '');
+
+$GLOBALS['TL_LANG']['tl_kongresse']['aktiv_legend'] = 'Publication';
+$GLOBALS['TL_LANG']['tl_kongresse']['aktiv'] = array('Published', 'Publish the record');